@@ -1,14 +1,15 @@
 {
   "name": "angular-ts-tdd",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "description": "",
-  "main": "./server.js",
+  "types": "module",
+  "main": "./server.mjs",
   "bin": {
-    "angular-ts-tdd-server": "./server.js",
-    "angular-ts-tdd-select": "./select.js"
+    "angular-ts-tdd-server": "./server.mjs",
+    "angular-ts-tdd-select": "./select.mjs"
   },
   "scripts": {
-    "start": "karma start server/karma/karma.config.js"
+    "start": "karma start server/karma/karma.config.cjs"
   },
   "author": "",
   "license": "ISC",
@@ -17,30 +18,35 @@
     "url": "git+https://github.com/Pmyl/angular-ts-tdd.git"
   },
   "dependencies": {
+    "@angular/compiler-cli": "^13.2.2",
     "@types/node": "7.0.7",
     "angular2-template-loader": "^0.6.2",
+    "babel-loader": "^8.2.3",
+    "core-js": "^3.21.0",
+    "css-loader": "^5.0.0",
     "file-loader": "^1.1.11",
     "fork-ts-checker-webpack-plugin": "^5.1.0",
     "html-loader": "^0.5.5",
     "jasmine-core": "^2.99.1",
-    "karma": "^4.0.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-matchers": "^4.0.1",
+    "karma": "^6.3.16",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-jasmine": "^4.0.1",
+    "karma-jasmine-matchers": "^5.0.0",
     "karma-jasmine-spy-when": "^1.0.5",
     "karma-sourcemap-loader": "^0.3.8",
+    "karma-webpack": "^5.0.0",
+    "loader-utils": "^1.2.3",
+    "node-sass": "^7.0.1",
+    "sass-loader": "^12.4.0",
+    "to-string-loader": "^1.1.6",
+    "ts-loader": "^9.2.6",
+    "ts-node": "^10.5.0",
     "tsconfig-paths-webpack-plugin": "^3.3.0",
     "url-loader": "^4.1.0",
-    "node-sass": "^4.14.1",
-    "css-loader": "^5.0.0",
-    "karma-webpack": "^4.0.2",
-    "loader-utils": "^1.2.3",
-    "to-string-loader": "^1.1.6",
-    "sass-loader": "^7.1.0",
-    "ts-loader": "^8.0.1",
-    "webpack": "^4.29.3"
+    "webpack": "^5.68.0",
+    "zone.js": "^0.11.4"
   },
   "peerDependencies": {
-    "typescript": "^3.0.1"
+    "typescript": "^4.5.5"
   }
 }
